add MyBinder at the notebook link
When looking at the readme file through MyBinder, there are issues linking to local files, because they are filled out to be links to GitHub, not local to MyBinder. This prevents any link opening local notebooks in Jupyter, as it will go to GitHub. The best current alternate is to have a separate MyBinder link for each binder in the readme file.